==> 02_get_airport_location.ipynb <==
# Get airport location information from [fallingrain.com](http://www.fallingrain.com/world/index.html)

The location of the nearest airport is needed to scrape the weather information (see notebook 4) and may also be useful as a feature in the model itself.

import pandas as pd
import numpy as np
import requests
import re
from StringIO import StringIO

from selenium import webdriver
from selenium.webdriver.chrome.options import Options
from bs4 import BeautifulSoup

from csv_pkl_sql import save_it

## Import cleaned location data

location_key = pd.read_pickle('../pkl/00_cleaned_city_names.pkl')

location_key.head(2)

location_key.isnull().sum()

data_countries = location_key.country.unique()
data_countries

Scrape a table of country abbreviations from fallingrain.com that will be used to build the URLs for country information.

tables = pd.read_html(requests.get('http://www.fallingrain.com/world/index.html').text)
countries = pd.DataFrame({'full':tables[0].values.ravel()}).dropna()
countries[['abbrev', 'name']] = (countries.full
                                 .str.extract(r"""([A-Z]{2}) (.+)""", expand=True)
                                 )

mask = countries.name.isin(data_countries)
countries = countries[mask].reset_index(drop=True)
assert mask.sum() == len(data_countries)

countries['url'] = countries.abbrev.apply(lambda x: 'http://www.fallingrain.com/world/{}/'.format(x))

United States state abbreviations

state_abbreviations = pd.read_csv(StringIO(requests.get('http://www.fonz.net/blog/wp-content/uploads/2008/04/states.csv').text))

## Scrape the airports

### All countries except the United States

countries['airports'] = countries.loc[countries.abbrev!='US','url'].apply(lambda x: x+'airports.html')

df_list = list()
for url in countries.airports:
    if url is not np.NaN:
        table = pd.read_html(url)[0]
        table.columns = table.iloc[0]
        table = table.iloc[1:]
        
        country_code = re.search(r"""world\/([A-Z]{2})\/""",url).group(1)
        country_name = countries.loc[countries.abbrev==country_code, 'name'].values[0]
        table['country'] = country_name
        df_list.append(table)
        
        
airports_df_1 = pd.concat(df_list, axis=0)

### Just the United States

states = pd.DataFrame({'states':location_key.loc[location_key.country=='United States','province'].unique()})
states = pd.merge(states, state_abbreviations, left_on='states', right_on='State', how='left')

states.loc[states.states=='Puerto Rico', 'Abbreviation'] = 'PR'
states.loc[states.states=='American Samoa', 'Abbreviation'] = 'AS'
states.loc[states.states=='Virgin Islands', 'Abbreviation'] = 'VI'
states.drop(['State'], axis=1, inplace=True)

states['airports'] = states.Abbreviation.apply(lambda x: 'http://www.fallingrain.com/world/US/{}/airports.html'.format(x))

df_list = list()
for url in states.drop_duplicates(subset=['Abbreviation']).airports:
    if url is not np.NaN:
        table = pd.read_html(url)[0]
        table.columns = table.iloc[0]
        table = table.iloc[1:]
        
        state_code = re.search(r"""world\/US\/([A-Z]{2})\/""",url).group(1)
        state_name = states.loc[states.Abbreviation==state_code, 'states'].values[0]
        table['state'] = state_name
        table['country'] = 'United States'
        df_list.append(table)
        
airports_df_2 = pd.concat(df_list, axis=0)

## Combine airport dataframes
Combine and clean columns.

airports = pd.concat([airports_df_1, airports_df_2], axis=0).reset_index(drop=True)

# Clean up column names
name_mapper = dict([(x,x.lower().replace(' ','_')) 
               for x in ['Kind','City','Name','Latitude','Longitude','Max Runway']])

airports = airports.rename(columns=name_mapper)

# Column formatting
airports['latitude'] = airports.latitude.str.replace(r"""\([NS]\)""", '').astype(float)
airports['longitude'] = airports.longitude.str.replace(r"""\([EW]\)""", '').astype(float)
airports['max_runway'] = airports.max_runway.str.replace(r""" ft""", '').astype(float)

# Extract just the medium and large airports
mask = airports.kind.isin(['Medium','Large'])
airports = airports[mask]

airports.head()

# Unused attempt at using airport name and location fuzzy string matching to determine latitude and longitude
# Decided to use google search instead

# airport_array = arg_air['name'].str.lower().str.replace(' airport','')
# arg_loc['airport_index'] = -1

# for row,dat in arg_loc.iterrows():
#     location_text = dat[dat.location_type].lower()
#     location_index = dat.name
    
#     match_value = process.extractOne(location_text, airport_array)[0]
#     match_index = airport_array.loc[airport_array==match_value].index[0]
    
#     arg_loc.loc[location_index, 'airport_index'] = match_index
#     print(location_text, match_value)


airports.shape[0]

save_it(airports, '02_airport_information_fallingrain')
