==> 04_scrape_weekly_weather_data.ipynb <==
# Scrape weather data based on date and location

The historical, international weather data on Weather Underground is keyed by airport abbreviation in the url. So, for each outbreak location, the closed airport is needed. To key these two sets of information, latitude and longitude are used.

import pandas as pd
import numpy as np
import re
import time
import dill
from datetime import timedelta
from csv_pkl_sql import save_it, pkl_it

## Load airport and latitude/longitude data

lat_long_data = pd.read_pickle('../pkl/01_latitude_longitude_google.pkl')
lat_long_data.head(1)

airport_info = pd.read_pickle('../pkl/02_airport_information_fallingrain.pkl')
airport_info.head(1)

The approximation for closest airport is crude, given that it doesn't convert latitude and longitude to distance but rather uses them directly. Given the relatively short distances involved, I think this is fine for a first pass of this project.

airport_coords = airport_info[['latitude', 'longitude']].values[np.newaxis, :]
places_coords = np.rollaxis(lat_long_data[['latitude','longitude']].values[np.newaxis, :], 0, -1)

dist_coords = ((places_coords - airport_coords)**2).sum(axis=-1)
min_coords = dist_coords.argmin(axis=1)

print airport_coords.shape, places_coords.shape, dist_coords.shape, min_coords.shape

# Transfer the coordinates to the latitude/longitude data
merge_data = lat_long_data.copy()

print merge_data.shape

merge_data['airport_index'] = airport_info.index[min_coords]

# Now grap the airport and location info
df = airport_info.loc[merge_data.airport_index, ['country','name','FAA','IATA','ICAO']]
merge_data[['country','name','FAA','IATA','ICAO']] = df.set_index(merge_data.index)

print merge_data.shape

merge_data.head()

pkl_it(merge_data, '04_merged_latitude_longitude_airport_checkpoint')

## Add infection dates
Now combine with infection date data.

infection_data = pd.read_pickle('../pkl/03_infection_data_initial_import.pkl')
infection_data = infection_data[['date','location']]
infection_data.head(1)

print infection_data.shape, merge_data.shape

merge_all = pd.merge(infection_data, 
                     merge_data[['location','country','FAA','IATA','ICAO']], 
                     on='location', 
                     how='left').drop_duplicates()

print merge_all.shape

merge_all.head()

## Scrape data from Weather Underground

Now scrape from weather underground. I want time shifted data, so need to get one and two weeks beforehand.

weather_scrape = (merge_all[['date','country','IATA','ICAO']]
                  .drop_duplicates()
                  .set_index(['country','IATA','ICAO'])
                  )

weather_scrape['date1'] = weather_scrape.date - timedelta(days=7)
weather_scrape['date2'] = weather_scrape.date - timedelta(days=14)

weather_scrape = (weather_scrape
                  .stack()
                  .reset_index(level=-1, drop=True)
                  .reset_index()
                  .rename(columns={0:'date'})
                  .dropna(subset=['IATA','ICAO'], how='all')
                 )

weather_scrape.shape

def scrape_weekly_weather(date, df_row):
    # Scrape the weekly data table
    url_fmt = 'https://www.wunderground.com/history/airport/{}/{}/{}/{}/WeeklyHistory.html'
    
    try:
        url = url_fmt.format(df_row.ICAO, date.year, 
                             date.month, date.day)
    except:
        url = url_fmt.format(df_row.IATA, date.year, 
                             date.month, date.day)
    
    try:
        table = pd.read_html(url)[0].dropna(subset=['Max','Avg','Min','Sum'], how='all')
        table.columns = ['Measurement','Max','Avg','Min','Sum']
        table.set_index('Measurement', inplace=True)
        table = table.stack()
        time.sleep(1.0)
    except:
        table = pd.Series({'NULL':np.NaN}, index=pd.Index([0]))
    
    return table

# Broke up into sections and did this on three different computers to speed it up
# date_list is a divided up dataframe of weather_scrape

for ndate, date in enumerate(date_list):
    
    print ndate
    df_list = list()
    
    for num,(row,dat) in enumerate(airport_list.iterrows()):
        
        try:
            df = scrape_weekly_weather(date, dat)
        except:
            df = pd.Series({'NULL':np.NaN}, index=pd.Index([row]))

        df_list.append((date, dat.name, df))
        
    with open('../pkl/04_scrape_weekly_weather_data/df_list{}.pkl'.format(ndate),'w') as fh:
        dill.dump(df_list, fh)


def clean_weather_data(entry):
    index = pd.MultiIndex.from_tuples([(entry[0],
                                        entry[1])]*len(entry[2]),
                                      names=['date','index'])
    
    df = pd.DataFrame(entry[2].reset_index().values, 
                      index=index, 
                      columns=['measurement','type','value'])

    mask = (df.measurement.isin(['Max Temperature','Mean Temperature',
                                   'Min Temperature','Dew Point','Precipitation','Wind']))
    df = df.loc[mask]
    
    mask = ((((df.measurement=='Precipitation')&(df.type=='Sum'))|(df.type=='Avg')) & 
            ((df.measurement=='Precipitation')&(df.type=='Avg')).pipe(np.invert))
    df = df.loc[mask].drop(['type'], axis=1)
    
    df['value'] = (df.value
                   .str.replace('-', '')
                   .str.extract(r"""([0-9.-]+)""", expand=True)
                   .astype(float)
                   )
    
    return df

df_clean = list()


for i in range(134):
    with open('../pkl/04_scrape_weekly_weather_data/df_list{}.pkl'.format(i), 'r') as fh:
        df_list = dill.load(fh)
    
    for df in enumerate(df_list):
        if not df[1][2].isnull().all():
            df_clean.append(clean_weather_data(df[1]))

weather_combined = pd.concat(df_clean, axis=0)
weather_combined.head()

weather_combined = pd.merge(weather_combined.reset_index(level=-1), 
                            airport_list, 
                            left_on='index', 
                            right_index=True).drop(['index'], axis=1).reset_index()

## Shift historical weather data

def time_shift(df, feature, week=1):
    new_df = (pd.merge(df[['date', feature]].reset_index(),
                       df[['date'+str(week), feature]].reset_index(),
                       left_on=df.index.names + ['date'], 
                       right_on=df.index.names + ['date'+str(week)],
                       suffixes=('',str(week)), 
                       how='inner')
              .drop(['date'+str(week)] + df.index.names, axis=1)
              .reset_index(level=-1, drop=True))
        
    return new_df


def create_weather_feature(df, feature):
    df_new = (df.loc[df.measurement==feature]
             .set_index(['ICAO','IATA','date','measurement'])
             .unstack())
    
    df_new = df_new.reset_index(level=-1)
    df_new.columns = ['date', feature]

    df_new['date1'] = df_new.date + timedelta(days=7)
    df_new['date2'] = df_new.date + timedelta(days=14)

    df_new1 = (df_new
            .groupby(level=[0,1])
            .apply(lambda x: time_shift(x,feature, 1))
            .reset_index(level=-1,drop=True))
    
    df_new2 = (df_new
            .groupby(level=[0,1])
            .apply(lambda x: time_shift(x, feature, 2))
            .reset_index(level=-1,drop=True))
    
    df_new = pd.merge(df_new1.reset_index(),
                      df_new2.reset_index().drop([feature], axis=1),
                      on=df_new1.index.names + ['date']).set_index(df_new1.index.names)
    
    return df_new

# Shift the one and two week prior data

max_temp = create_weather_feature(weather_combined, 'Max Temperature').set_index('date',append=True)
mean_temp = create_weather_feature(weather_combined, 'Mean Temperature').set_index('date',append=True)
min_temp = create_weather_feature(weather_combined, 'Min Temperature').set_index('date',append=True)
dew_point = create_weather_feature(weather_combined, 'Dew Point').set_index('date',append=True)
precipitation = create_weather_feature(weather_combined, 'Precipitation').set_index('date',append=True)
wind = create_weather_feature(weather_combined, 'Wind').set_index('date',append=True)

[x.shape for x in max_temp, mean_temp, min_temp, dew_point, precipitation, wind]

[x.isnull().sum().max() for x in max_temp, mean_temp, min_temp, dew_point, precipitation, wind]

Impute missing data.

max_temp = max_temp.interpolate(method='linear', limit_direction='both')
mean_temp = mean_temp.interpolate(method='linear', limit_direction='both')
min_temp = min_temp.interpolate(method='linear', limit_direction='both')
dew_point = dew_point.interpolate(method='linear', limit_direction='both')
precipitation = precipitation.interpolate(method='linear', limit_direction='both')
wind = wind.interpolate(method='linear', limit_direction='both')

[x.shape for x in max_temp, mean_temp, min_temp, dew_point, precipitation, wind]

[x.isnull().sum().max() for x in max_temp, mean_temp, min_temp, dew_point, precipitation, wind]

Add location key.

airport = pd.read_pickle('../pkl/04_merged_latitude_longitude_airport_checkpoint.pkl')
airport.head(1)

max_temp = pd.merge(max_temp.reset_index(),
         airport[['ICAO','IATA','location']],
         on=['ICAO','IATA'],
         how='left')#.drop_duplicates(subset=['location'])

mean_temp = pd.merge(mean_temp.reset_index(),
         airport[['ICAO','IATA','location']],
         on=['ICAO','IATA'],
         how='left')


min_temp = pd.merge(min_temp.reset_index(),
         airport[['ICAO','IATA','location']],
         on=['ICAO','IATA'],
         how='left')


dew_point = pd.merge(dew_point.reset_index(),
         airport[['ICAO','IATA','location']],
         on=['ICAO','IATA'],
         how='left')

precipitation = pd.merge(precipitation.reset_index(),
         airport[['ICAO','IATA','location']],
         on=['ICAO','IATA'],
         how='left')

wind = pd.merge(wind.reset_index(),
         airport[['ICAO','IATA','location']],
         on=['ICAO','IATA'],
         how='left')

[x.shape for x in max_temp, mean_temp, min_temp, dew_point, precipitation, wind]

[x.isnull().sum().max() for x in max_temp, mean_temp, min_temp, dew_point, precipitation, wind]

max_temp = max_temp.drop(['ICAO','IATA'], axis=1).drop_duplicates(subset=['location','date'])
max_temp.columns = [x.lower().replace(' ', '_').replace('erature','') for x in max_temp.columns]

mean_temp = mean_temp.drop(['ICAO','IATA'], axis=1).drop_duplicates(subset=['location','date'])
mean_temp.columns = [x.lower().replace(' ', '_').replace('erature','') for x in mean_temp.columns]

min_temp = min_temp.drop(['ICAO','IATA'], axis=1).drop_duplicates(subset=['location','date'])
min_temp.columns = [x.lower().replace(' ', '_').replace('erature','') for x in min_temp.columns]

dew_point = dew_point.drop(['ICAO','IATA'], axis=1).drop_duplicates(subset=['location','date'])
dew_point.columns = [x.lower().replace(' ', '_').replace('erature','') for x in dew_point.columns]

precipitation = precipitation.drop(['ICAO','IATA'], axis=1).drop_duplicates(subset=['location','date'])
precipitation.columns = [x.lower().replace(' ', '_').replace('erature','') for x in precipitation.columns]

wind = wind.drop(['ICAO','IATA'], axis=1).drop_duplicates(subset=['location','date'])
wind.columns = [x.lower().replace(' ', '_').replace('erature','') for x in wind.columns]

[x.shape for x in max_temp, mean_temp, min_temp, dew_point, precipitation, wind]

[x.isnull().sum().max() for x in max_temp, mean_temp, min_temp, dew_point, precipitation, wind]

print max_temp.shape

weather_final = pd.merge(max_temp, mean_temp, on=['date','location'], how='inner')
weather_final = pd.merge(weather_final, min_temp, on=['date','location'], how='inner')
weather_final = pd.merge(weather_final, dew_point, on=['date','location'], how='inner')
weather_final = pd.merge(weather_final, precipitation, on=['date','location'], how='inner')
weather_final = pd.merge(weather_final, wind, on=['date','location'], how='inner')

print weather_final.shape

weather_final.isnull().sum().max()

weather_final.head(1).T

weather_final.dtypes

save_it(weather_final, '04_weekly_weather')
