==> 06_get_population_density.ipynb <==
# Get population density

Population density for the outbreak region used the [NASA Socioeconomic Data and Applications Center (SEDAC)](http://sedac.ciesin.columbia.edu/data/set/gpw-v4-population-density) Gridded Population and Population Density Map of the World. This is a `tif` file with population densities encoded in the file. 

Extraction of the relevant data requires `osgeo` which would only run in a separate python installation. Thus, the python code below is encapsulated in a script so that Jupyter did not also have to be installed in the osgeo-specific installation.

! cat 06_get_population_density.py

! ./06_get_population_density.py
